+++ carp_mobile_sensing/lib/sampling_packages/device/device.g.dart
@@ -17,6 +17,9 @@ DeviceDatum _$DeviceDatumFromJson(Map<String, dynamic> json) {
     hardware: json['hardware'] as String,
   )
     ..id = json['id'] as String
+    ..timestamp = json['timestamp'] == null
+        ? null
+        : DateTime.parse(json['timestamp'] as String)
     ..sdk = json['sdk'] as String
     ..release = json['release'] as String;
 }
@@ -31,6 +34,7 @@ Map<String, dynamic> _$DeviceDatumToJson(DeviceDatum instance) {
   }
 
   writeNotNull('id', instance.id);
+  writeNotNull('timestamp', instance.timestamp?.toIso8601String());
   writeNotNull('platform', instance.platform);
   writeNotNull('device_id', instance.deviceId);
   writeNotNull('hardware', instance.hardware);
@@ -46,6 +50,9 @@ Map<String, dynamic> _$DeviceDatumToJson(DeviceDatum instance) {
 BatteryDatum _$BatteryDatumFromJson(Map<String, dynamic> json) {
   return BatteryDatum()
     ..id = json['id'] as String
+    ..timestamp = json['timestamp'] == null
+        ? null
+        : DateTime.parse(json['timestamp'] as String)
     ..batteryLevel = json['battery_level'] as int
     ..batteryStatus = json['battery_status'] as String;
 }
@@ -60,6 +67,7 @@ Map<String, dynamic> _$BatteryDatumToJson(BatteryDatum instance) {
   }
 
   writeNotNull('id', instance.id);
+  writeNotNull('timestamp', instance.timestamp?.toIso8601String());
   writeNotNull('battery_level', instance.batteryLevel);
   writeNotNull('battery_status', instance.batteryStatus);
   return val;
@@ -68,6 +76,9 @@ Map<String, dynamic> _$BatteryDatumToJson(BatteryDatum instance) {
 FreeMemoryDatum _$FreeMemoryDatumFromJson(Map<String, dynamic> json) {
   return FreeMemoryDatum()
     ..id = json['id'] as String
+    ..timestamp = json['timestamp'] == null
+        ? null
+        : DateTime.parse(json['timestamp'] as String)
     ..freePhysicalMemory = json['free_physical_memory'] as int
     ..freeVirtualMemory = json['free_virtual_memory'] as int;
 }
@@ -82,6 +93,7 @@ Map<String, dynamic> _$FreeMemoryDatumToJson(FreeMemoryDatum instance) {
   }
 
   writeNotNull('id', instance.id);
+  writeNotNull('timestamp', instance.timestamp?.toIso8601String());
   writeNotNull('free_physical_memory', instance.freePhysicalMemory);
   writeNotNull('free_virtual_memory', instance.freeVirtualMemory);
   return val;
@@ -90,6 +102,9 @@ Map<String, dynamic> _$FreeMemoryDatumToJson(FreeMemoryDatum instance) {
 ScreenDatum _$ScreenDatumFromJson(Map<String, dynamic> json) {
   return ScreenDatum()
     ..id = json['id'] as String
+    ..timestamp = json['timestamp'] == null
+        ? null
+        : DateTime.parse(json['timestamp'] as String)
     ..screenEvent = json['screen_event'] as String;
 }
 
@@ -103,6 +118,7 @@ Map<String, dynamic> _$ScreenDatumToJson(ScreenDatum instance) {
   }
 
   writeNotNull('id', instance.id);
+  writeNotNull('timestamp', instance.timestamp?.toIso8601String());
   writeNotNull('screen_event', instance.screenEvent);
   return val;
 }