+++ carp_mobile_sensing/lib/sampling_packages/sensors/pedometer_probe.dart
@@ -14,6 +14,6 @@ part of sensors;
 ///
 /// Note that the [Pedometer] plugin returns the steps taken since last system boot.
 class PedometerProbe extends StreamProbe {
-  Stream<DataPoint> get stream => Pedometer.stepCountStream.map(
-      (StepCount count) => DataPoint.fromData(PedometerDatum(count.steps)));
+  Stream<Datum> get stream => Pedometer.stepCountStream
+      .map((StepCount count) => PedometerDatum(count.steps));
 }