+++ carp_mobile_sensing/lib/sampling_packages/sensors/sensor_probes.dart
@@ -11,15 +11,11 @@ part of sensors;
 /// sensor probes.
 abstract class BufferingSensorProbe extends BufferingPeriodicStreamProbe {
   MultiDatum datum = MultiDatum();
-  DateTime samplingStart;
 
-  Future<DataPoint> getDataPoint() async => DataPoint.fromData(datum)
-    ..carpHeader.startTime = samplingStart
-    ..carpHeader.endTime = DateTime.now();
+  Future<Datum> getDatum() async => datum;
 
   void onSamplingStart() {
     datum = MultiDatum();
-    samplingStart = DateTime.now();
   }
 
   void onSamplingEnd() {}
@@ -30,8 +26,8 @@ abstract class BufferingSensorProbe extends BufferingPeriodicStreamProbe {
 /// Note that this probe generates a lot of data and should be used
 /// with caution.
 class AccelerometerProbe extends StreamProbe {
-  Stream<DataPoint> get stream => accelerometerEvents.map((event) =>
-      DataPoint.fromData(AccelerometerDatum.fromAccelerometerEvent(event)));
+  Stream<Datum> get stream => accelerometerEvents
+      .map((event) => AccelerometerDatum.fromAccelerometerEvent(event));
 }
 
 /// A probe that collects accelerometer events and buffers them and return
@@ -50,8 +46,8 @@ class BufferingAccelerometerProbe extends BufferingSensorProbe {
 /// Note that this probe generates a lot of data and should be used
 /// with caution.
 class GyroscopeProbe extends StreamProbe {
-  Stream<DataPoint> get stream => gyroscopeEvents.map(
-      (event) => DataPoint.fromData(GyroscopeDatum.fromGyroscopeEvent(event)));
+  Stream<Datum> get stream =>
+      gyroscopeEvents.map((event) => GyroscopeDatum.fromGyroscopeEvent(event));
 }
 
 /// A probe that collects gyroscope events and buffers them and return