+++ carp_mobile_sensing/lib/sampling_packages/sensors/sensors.g.dart
@@ -11,7 +11,11 @@ AccelerometerDatum _$AccelerometerDatumFromJson(Map<String, dynamic> json) {
     x: (json['x'] as num)?.toDouble(),
     y: (json['y'] as num)?.toDouble(),
     z: (json['z'] as num)?.toDouble(),
-  )..id = json['id'] as String;
+  )
+    ..id = json['id'] as String
+    ..timestamp = json['timestamp'] == null
+        ? null
+        : DateTime.parse(json['timestamp'] as String);
 }
 
 Map<String, dynamic> _$AccelerometerDatumToJson(AccelerometerDatum instance) {
@@ -24,6 +28,7 @@ Map<String, dynamic> _$AccelerometerDatumToJson(AccelerometerDatum instance) {
   }
 
   writeNotNull('id', instance.id);
+  writeNotNull('timestamp', instance.timestamp?.toIso8601String());
   writeNotNull('x', instance.x);
   writeNotNull('y', instance.y);
   writeNotNull('z', instance.z);
@@ -35,7 +40,11 @@ GyroscopeDatum _$GyroscopeDatumFromJson(Map<String, dynamic> json) {
     x: (json['x'] as num)?.toDouble(),
     y: (json['y'] as num)?.toDouble(),
     z: (json['z'] as num)?.toDouble(),
-  )..id = json['id'] as String;
+  )
+    ..id = json['id'] as String
+    ..timestamp = json['timestamp'] == null
+        ? null
+        : DateTime.parse(json['timestamp'] as String);
 }
 
 Map<String, dynamic> _$GyroscopeDatumToJson(GyroscopeDatum instance) {
@@ -48,6 +57,7 @@ Map<String, dynamic> _$GyroscopeDatumToJson(GyroscopeDatum instance) {
   }
 
   writeNotNull('id', instance.id);
+  writeNotNull('timestamp', instance.timestamp?.toIso8601String());
   val['x'] = instance.x;
   val['y'] = instance.y;
   val['z'] = instance.z;
@@ -60,7 +70,11 @@ LightDatum _$LightDatumFromJson(Map<String, dynamic> json) {
     stdLux: json['std_lux'] as num,
     minLux: json['min_lux'] as num,
     maxLux: json['max_lux'] as num,
-  )..id = json['id'] as String;
+  )
+    ..id = json['id'] as String
+    ..timestamp = json['timestamp'] == null
+        ? null
+        : DateTime.parse(json['timestamp'] as String);
 }
 
 Map<String, dynamic> _$LightDatumToJson(LightDatum instance) {
@@ -73,6 +87,7 @@ Map<String, dynamic> _$LightDatumToJson(LightDatum instance) {
   }
 
   writeNotNull('id', instance.id);
+  writeNotNull('timestamp', instance.timestamp?.toIso8601String());
   writeNotNull('mean_lux', instance.meanLux);
   writeNotNull('std_lux', instance.stdLux);
   writeNotNull('min_lux', instance.minLux);
@@ -83,7 +98,11 @@ Map<String, dynamic> _$LightDatumToJson(LightDatum instance) {
 PedometerDatum _$PedometerDatumFromJson(Map<String, dynamic> json) {
   return PedometerDatum(
     json['step_count'] as int,
-  )..id = json['id'] as String;
+  )
+    ..id = json['id'] as String
+    ..timestamp = json['timestamp'] == null
+        ? null
+        : DateTime.parse(json['timestamp'] as String);
 }
 
 Map<String, dynamic> _$PedometerDatumToJson(PedometerDatum instance) {
@@ -96,6 +115,7 @@ Map<String, dynamic> _$PedometerDatumToJson(PedometerDatum instance) {
   }
 
   writeNotNull('id', instance.id);
+  writeNotNull('timestamp', instance.timestamp?.toIso8601String());
   writeNotNull('step_count', instance.stepCount);
   return val;
 }